refactor: 💡 PR Corrige les retours design sur la liste articles
✅ Closes: #268
