+++ front/components/StyledText.tsx
@@ -1,56 +1,9 @@
 import type { FC } from "react";
 import * as React from "react";
-import type { StyleProp, TextStyle } from "react-native";
-import { StyleSheet } from "react-native";
 
-import { FontWeight } from "../constants/Layout";
+import { FontNames, getFontFromWeight } from "../constants";
 import type { TextProps } from "./Themed";
 import { Text } from "./Themed";
-
-export enum FontNames {
-  comfortaa = "comfortaa",
-  spaceMono = "space-mono",
-  avenir = "avenir",
-}
-const fontsMap = {
-  [FontWeight.thin]: "light",
-  [FontWeight.extralight]: "light",
-  [FontWeight.light]: "light",
-  [FontWeight.normal]: "regular",
-  [FontWeight.medium]: "medium",
-  [FontWeight.semibold]: "semibold",
-  [FontWeight.bold]: "bold",
-  [FontWeight.extrabold]: "bold",
-  [FontWeight.black]: "black",
-  bold: "bold",
-  normal: "regular",
-};
-type FontWeightValues =
-  | "100"
-  | "200"
-  | "300"
-  | "400"
-  | "500"
-  | "600"
-  | "700"
-  | "800"
-  | "900"
-  | "bold"
-  | "normal";
-export const getFontFamilyName = (
-  fontName: string,
-  fontWeight: FontWeightValues
-): string => `${fontName}-${fontsMap[fontWeight]}`;
-
-const getFontFromWeight = (fontName: string, style?: StyleProp<TextStyle>) => {
-  const fontWeight = style
-    ? StyleSheet.flatten(style).fontWeight ?? FontWeight.normal
-    : FontWeight.normal;
-  return {
-    fontFamily: getFontFamilyName(fontName, fontWeight),
-    fontWeight: FontWeight.normal,
-  };
-};
 
 export const MonoText: FC<TextProps> = (props) => {
   return (
