+++ front/components/modal.component.tsx
@@ -9,7 +9,7 @@ import { WebView } from "react-native-webview";
 
 import { FontWeight, Paddings } from "../constants";
 import { AssestUtils } from "../utils";
-import Button from "./form/Button";
+import Button from "./form/button.component";
 import { View } from "./Themed";
 
 interface Props {
