+++ front/navigation/index.tsx
@@ -15,9 +15,7 @@ import AppLogo from "../assets/images/logo.svg";
 import { Backdrop, Icomoon, IcomoonIcons } from "../components";
 import Menu from "../components/menu.component";
 import { Text, View } from "../components/Themed";
-import { Paddings, Sizes } from "../constants";
-import Colors from "../constants/Colors";
-import Labels from "../constants/Labels";
+import { Colors, Labels, Paddings, Sizes } from "../constants";
 import ConditionsOfUse from "../screens/ConditionsOfUse";
 import LegalNotice from "../screens/LegalNotice";
 import LoadingScreen from "../screens/LoadingScreen";
