+++ front/screens/epdsSurvey/epdsLoadPreviousSurvey.component.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { StyleSheet } from "react-native";
 
-import Button from "../../components/form/Button";
+import Button from "../../components/form/button.component";
 import { CommonText } from "../../components/StyledText";
 import { View } from "../../components/Themed";
 import { Colors, FontWeight, Labels, Paddings, Sizes } from "../../constants";
