+++ front/screens/epdsSurvey/epdsSurveyFooter.component.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { StyleSheet } from "react-native";
 import type SwiperFlatListRefProps from "react-native-swiper-flatlist";
 
-import Button from "../../components/form/Button";
+import Button from "../../components/form/button.component";
 import Icomoon, { IcomoonIcons } from "../../components/icomoon.component";
 import { View } from "../../components/Themed";
 import { Colors, Labels, Paddings } from "../../constants";
