+++ front/screens/index.ts
@@ -3,7 +3,7 @@ import ArticleDetail from "./ArticleDetail";
 import ConditionsOfUse from "./ConditionsOfUse";
 import EpdsSurveyScreen from "./epdsSurvey/epdsSurveyScreen.component";
 import LegalNotice from "./LegalNotice";
-import ListArticles from "./ListArticles";
+import ListArticles from "./listArticles.component";
 import TabCalendarScreen from "./TabCalendarScreen";
 import TabFavoritesScreen from "./TabFavoritesScreen";
 import TabHomeScreen from "./TabHomeScreen";
