+++ front/screens/listArticles.component.tsx
